+++ src/utils/musicSdk/kw/util.js
@@ -1,5 +1,7 @@
 // import BackgroundTimer from 'react-native-background-timer'
 // import { httpGet, httpFetch } from '../../request'
+import { toMD5 } from '../utils'
+import { aesEncryptSync, aesDecryptSync, AES_MODE } from '@/utils/nativeModules/crypto'
 export { default as decodeLyric } from './decodeLyric'
 
 // const kw_token = {
@@ -175,5 +177,47 @@ export const lrcTools = {
     if (tools.tags.length) lrcs = `${tools.tags.join('\n')}\n${lrcs}`
     // console.log(lrcs)
     return lrcs
+  },
+}
+
+
+// const createAesEncrypt = (buffer, mode, key, iv) => {
+//   const cipher = createCipheriv(mode, key, iv)
+//   return Buffer.concat([cipher.update(buffer), cipher.final()])
+// }
+
+// const createAesDecrypt = (buffer, mode, key, iv) => {
+//   const cipher = createDecipheriv(mode, key, iv)
+//   return Buffer.concat([cipher.update(buffer), cipher.final()])
+// }
+
+export const wbdCrypto = {
+  aesMode: 'aes-128-ecb',
+  // aesKey: Buffer.from([112, 87, 39, 61, 199, 250, 41, 191, 57, 68, 45, 114, 221, 94, 140, 228], 'binary'),
+  aesKey: 'cFcnPcf6Kb85RC1y3V6M5A==',
+  aesIv: '',
+  appId: 'y67sprxhhpws',
+  decodeData(base64Result) {
+    // const data = Buffer.from(decodeURIComponent(base64Result), 'base64')
+    // return JSON.parse(createAesDecrypt(data, this.aesMode, this.aesKey, this.aesIv).toString())
+    const data = decodeURIComponent(base64Result)
+    return JSON.parse(aesDecryptSync(data, this.aesKey, this.aesIv, AES_MODE.ECB_128_NoPadding))
+  },
+  createSign(data, time) {
+    const str = `${this.appId}${data}${time}`
+    return toMD5(str).toUpperCase()
+  },
+  buildParam(jsonData) {
+    // const data = Buffer.from(JSON.stringify(jsonData))
+    // const time = Date.now()
+
+    // const encodeData = createAesEncrypt(data, this.aesMode, this.aesKey, this.aesIv).toString('base64')
+    const data = Buffer.from(JSON.stringify(jsonData)).toString('base64')
+    const time = Date.now()
+
+    const encodeData = aesEncryptSync(data, this.aesKey, this.aesIv, AES_MODE.ECB_128_NoPadding)
+    const sign = this.createSign(encodeData, time)
+
+    return `data=${encodeURIComponent(encodeData)}&time=${time}&appId=${this.appId}&sign=${sign}`
   },
 }