+++ .ncurc.js
@@ -1,6 +1,7 @@
 module.exports = {
   upgrade: true,
   reject: [
+    'react-native-navigation',
     '@types/react-native',
     'message2call',
     'react',