+++ publish/changeLog.md
@@ -2,9 +2,15 @@
 
 - 新增棕色主题“泥牛入海”
 
+### 优化
+
+- 不再丢弃kg源逐行歌词（@helloplhm-qwq）
+- 支持kw源排行榜显示大小（revert @Folltoshe #1460）
+
 ### 修复
 
 - 修复mg歌词在某些情况下获取失败的问题
+- 修复mg歌单搜索（@helloplhm-qwq）
 - 修复kg最新评论无法获取的问题（@helloplhm-qwq）
 
 ### 其他