+++ publish/changeLog.md
@@ -15,4 +15,4 @@
 
 ### 其他
 
-- 更新 React native 到 v0.73.4
+- 更新 React native 到 v0.73.5